move bare metal implementation to bare-metal/
