+++ bare-metal/uart.c
