+++ bare-metal/uart.h
