+++ bare-metal/ws2812_config.h
