+++ light_ws2812/ws2812_config.h
@@ -1,0 +1,21 @@
+/*
+ * light_ws2812_config.h
+ *
+ * Created: 18.01.2014 09:58:15
+ *
+ * User Configuration file for the light_ws2812_lib
+ *
+ */ 
+
+
+#ifndef WS2812_CONFIG_H_
+#define WS2812_CONFIG_H_
+
+///////////////////////////////////////////////////////////////////////
+// Define I/O pin
+///////////////////////////////////////////////////////////////////////
+
+#define ws2812_port D     // Data port
+#define ws2812_pin  7     // Data out pin
+
+#endif /* WS2812_CONFIG_H_ */
